groq_chat.py
from langchain_groq import ChatGroq
import os
from dotenv import load_dotenv

load_dotenv()

def load_llm():

    llm = ChatGroq(
        groq_api_key=os.getenv("GROQ_API_KEY"),
        model_name="llama-3.1-8b-instant",
        temperature=0
    )

    return llm