requirements.txt
streamlit
langchain
langchain-community
langchain-groq
faiss-cpu
python-dotenv
pypdf
sentence-transformers